Refactor data handling and layout providers
