+++ TextInsightsGrid.module.css
@@ -1,0 +1,18 @@
+.root {
+  width: 100%;
+}
+
+.category {
+  text-transform: uppercase;
+  letter-spacing: 0.04em;
+}
+
+.insightTitle {
+  font-weight: 600;
+  line-height: 1.3;
+}
+
+.description {
+  line-height: 1.6;
+}
+
