+++ generate_saleschart_data.py
@@ -2,7 +2,7 @@
 import pandas as pd
 import json
 
-df = pd.read_json("/home/ubuntu/cyber-dashboard-template-project/public/mocks/simulated_data.json")
+df = pd.read_json("/workspace/public/mocks/simulated_data.json")
 
 # Conta le occorrenze di ogni tipo di reciprocità
 reciprocity_counts = df["ReciprocityType"].value_counts().reset_index()
@@ -13,7 +13,7 @@ sales_chart_data = {
     "labels": reciprocity_counts["label"].tolist()
 }
 
-with open("/home/ubuntu/cyber-dashboard-template-project/public/mocks/SalesChart.json", "w") as f:
+with open("/workspace/public/mocks/SalesChart.json", "w") as f:
     json.dump(sales_chart_data, f, indent=4)
 
 print("Dati per SalesChart generati e salvati in SalesChart.json")
